+++ lntest/itest/list_off_test.go
@@ -5,5 +5,4 @@ package itest
 
 import "github.com/lightningnetwork/lnd/lntemp"
 
-// TODO(yy): remove the temp.
-var allTestCasesTemp = []*lntemp.TestCase{}
+var allTestCases = []*lntemp.TestCase{}
