+++ lntest/itest/list_on_test.go
@@ -5,8 +5,7 @@ package itest
 
 import "github.com/lightningnetwork/lnd/lntemp"
 
-// TODO(yy): remove the temp.
-var allTestCasesTemp = []*lntemp.TestCase{
+var allTestCases = []*lntemp.TestCase{
 	{
 		Name:     "update channel status",
 		TestFunc: testUpdateChanStatus,
