--- lntest/itest/lnd_test_list_on_test.go
@@ -1,6 +1,0 @@
-//go:build rpctest
-// +build rpctest
-
-package itest
-
-var allTestCases = []*testCase{}
